Tools installation should use go.mod for dependency versioning
**Desired Behavior**

* Installed tools versions are locked the versions of installed tools to specific versions
* Versions of the dependencies used by installed tools are controlled by `go.mod` and `go.sum`

**Actual behavior**

The Dockerfile includes tool installations locked to a specific commit by referencing that commit in the command `go install <package>@<commit>`:

https://github.com/GoogleContainerTools/kaniko/blob/24846d22b11f20edc03dff7ac3b3959019969873/deploy/Dockerfile#L25-L32

However, this form of the `go install` command bypasses the local `go.mod`
dependency versioning and also ignores dependencies in the `vendor` directory.  Instead it goes to the project commit referenced and downloads the versions of the dependencies listed in that external package and also uses those at runtime.  This documented but easily overlooked behavior.

See https://github.com/golang/go/issues/48332 for background.

**Proposed fix**

To address this we must add a file that imports the packages we
intend to install - it only needs to import them not do anything else with them.

With that done, manually calling `go install <package>@<commit>` will properly add `<package>` and its dependencies to the `go.mod` and `go.sum` files.
Then calling `go mod vendor` will update the `vendor` directory to include the updated dependency.
In both calls the dependency versions in `go.mod` will take precedence and can be modified if needed.

Finally, future scripted calls must be in the form `go install <package>` so that `<package>` and dependencies will be looked up using the local configuration.

**Side effects of proposed fix**

* Installed tools and their dependencies are are included in the `vendor` directory  - `Docker-slim` is less slim due the additional vendoring, but that scenario is not built by Verta.